Linear autoencoder
encoder and decoder made of one linear layer.
by encoder , a compressed representation is produced.
decoder is responsible for forming a reconstrcted image, this will increase the dimension of the compressed representation.
The final output layer will have a sigmoid activation function to scale its values to be gray scale pixel values;between zero and one. This is necessary in oder to make us abl to get values and match this with input values.