Convolutional autoencoder
In comparison to linear autoencoder, convolutional autoencoder reduce the blurriness but instead gives some blocky artifacts and struggle with curved areas. The reason is transpose convolutional tend to overlap and produce a grid like pattern.
A better solution is combining the nearest interpolation with normal convolutional layer . (the both ways are written in this file)