A Sentiment Analysis model by using AWS
Using SageMaker in AWS is a requirement of constructing a complete project . The goal of this project is to have a simple web page which a user can use to enter a movie review. The web page will then send the review off to the deployed model which will predict the sentiment of the entered review.